Update README.md
Added the description of my code design , classes, methods and other key details to the read me file. Moreover, I have also added the GitHub repository which contains my work for part 1 and part 2. This was the repository I was using to implement version control and protect form loss of work.